Fix build warnings (#572)
* Update spdlog to v1.17.0 to get rid of deprecation warnings when compiling on MacOS

* Use latest master commit of msgpack11 to fix MacOS build warnings and do not bundle it with the library anymore

* Fix unused variable preferTLS warning in MacOS

* Remove CLI11 and cpp-linenoise from third_party directory and fetch newer versions via ws CMakeLists - now there are no more dependencies stored in the repository

* Fix build warnings on Windows with clang 21

* Fix build warnings when building without zlib

* Remove reference to previously removed third_party whitespace script

* Update msgpack11 to latest commit

* Update cpp_linenoise to latest version

* Fix compilation warnings and errors in IXSocketMbedTLS.cpp

* Remove embedded Catch2 and use external repository with tag v3.15.1

* Fix build warnings in IXSocketOpenSSL.cpp

* Update GitHub Actions unittest_mac_tsan_openssl to openssl 3.5

+++ ixwebsocket/IXSocketServer.cpp
@@ -331,7 +331,7 @@ namespace ix
             socklen_t addressLen = sizeof(client);
             memset(&client, 0, sizeof(client));
 
-            if ((clientFd = accept(_serverFd, (struct sockaddr*) &client, &addressLen)) < 0)
+            if ((clientFd = static_cast<int>(accept(_serverFd, (struct sockaddr*) &client, &addressLen))) < 0)
             {
                 if (!Socket::isWaitNeeded())
                 {
